+++ CHANGES.md
@@ -1,6 +1,10 @@
 # CHANGES
 ## RosarioSIS Student Information System
 
+Changes in 5.3.1
+----------------
+- Fix Last row inputs for non admin users in Transcripts.fnc.php & ReportCards.fnc.php
+
 Changes in 5.3
 --------------
 - Fix SQL error more than one row returned by a subquery in Search.fnc.php & Preferences.php